requirements.txt
sounddevice==0.5.0
numpy==2.0.1
librosa==0.10.2.post1
joblib==1.4.2
pynput==1.7.7
pyserial==3.5
scipy==1.14.0
scikit-learn==1.5.1
seaborn==0.13.2
